feat: Add ability to create manual spans for datadog.
